requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-crispy-forms==1.6.1
django-map-widgets==0.1.9
idna==2.5
jsonfield==2.0.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
phonenumberslite==8.8.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
django-storages==1.5.2
apache-libcloud==2.0.0
gunicorn==19.7.1
dj-database-url==0.4.2
whitenoise==3.3.0

Babel==2.5.0
django-phonenumber-field==1.3.0
djangorestframework==3.6.4

